Move `logout` to `commoncommands`
